Silenced more warnings

+++ src/vio2sf/desmume/bios.c
@@ -308,7 +308,7 @@ u32 devide(armcpu_t* cpu)
      
      cpu->R[0] = (u32)(num / dnum);
      cpu->R[1] = (u32)(num % dnum);
-     cpu->R[3] = (u32) (((s32)cpu->R[0])<0 ? -cpu->R[0] : cpu->R[0]);
+     cpu->R[3] = (u32) (((s32)cpu->R[0])<0 ? -((s32)cpu->R[0]) : cpu->R[0]);
      
      return 6;
 }
